text_to_image/requirements.txt
celery==5.4.0
Django==5.0.7
django-celery-results==2.5.1
pillow==10.4.0
python-dotenv==1.0.1
redis==5.0.7
requests==2.32.3